up to 4.0.0

+++ DBService/Models/Interface/IDataManipulate.cs
@@ -8,15 +8,14 @@ namespace DBService.Models.Interface
 {
     public interface IDataManipulate
     {
-        string? GetSqlForDelete(int id);
+        string? GetSqlForDelete(long id);
         string? GetSqlForDeleteByKey(KeyValuePair<string, object?> criteria);
 
         string? GetSqlForInsert(IEnumerable<KeyValuePair<string, object?>> source);
-        string? GetSqlForUpdate(int id, IEnumerable<KeyValuePair<string, object?>> source);
+        string? GetSqlForUpdate(long id, IEnumerable<KeyValuePair<string, object?>> source);
         string? GetSqlForUpdateByKey(KeyValuePair<string, object?> criteria, IEnumerable<KeyValuePair<string, object?>> source);
         #region Multi Version
 
         #endregion
-
     }
 }
